changes to IndexT where required

+++ src/redecomp/RedecompMesh.cpp
@@ -273,7 +273,8 @@ void RedecompMesh::BuildRedecomp()
   r2p_elem_offsets_.reserve( n_ranks + 1 );
   r2p_elem_offsets_.resize( n_ranks + 1 );
   mpi_.SendRecvEach(
-      type<axom::Array<int>>(), [this]( int dest ) { return axom::Array<int>( { p2r_elems_.first[dest].size() } ); },
+      type<axom::Array<int>>(),
+      [this]( int dest ) { return axom::Array<int>( { static_cast<int>( p2r_elems_.first[dest].size() ) } ); },
       [this]( axom::Array<int>&& recv_data, int source ) { r2p_elem_offsets_[source + 1] = recv_data[0]; } );
   for ( int i{ 2 }; i < n_ranks + 1; ++i ) {
     r2p_elem_offsets_[i] += r2p_elem_offsets_[i - 1];
